move reference validation code to problem module
