+++ tests/test_types.py
@@ -5,8 +5,8 @@ from unittest.util import safe_repr
 
 from pydantic import ValidationError
 
-from algobattle.problem import InstanceModel
-from algobattle.util import AttributeReference, Role, SelfRef
+from algobattle.problem import InstanceModel, AttributeReference, SelfRef
+from algobattle.util import Role
 from algobattle.types import Ge, Interval, LaxComp, SizeIndex, UniqueItems
